Add purge_config option for all configs
Adding a purge_config option for all configs (specifically left out
services/lbaas.pp and services/vpnaas.pp as they require default
entries). Adding this option for ease of management and readability.
If users wish, they can use this to ensure that only the options they
are setting make it into the configs.
  *rebased my changes

Change-Id: I2286acb350ae6300f2e4c8dd320f2534ddf09bf9