Don't declare the service when service_enable is false
In environments using neutron-vpn-agent, the neutron-l3-agent package
may not be installed ( by setting package_ensure to false ). The
l3-agent service is always declared, even if service_enable is set to
false, which will cause puppet to try and check the status of the
service and generate failures because the init script is missing. This
commit makes the service declaration dependent on service_enable being
set to true, and modifies the spec tests to reflect this.

Change-Id: Ib223999ab3ed3e83f42e96ad91a0728681c377b2
Closes-bug: #1571697